python/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pytesseract==0.3.13
pillow==10.4.0
pymupdf==1.24.10
python-docx==1.1.2
openpyxl==3.1.5